Fixed problem with floating point test
By default floating-point literals have type 'double', not
'float'. When used in expressions with 'float' variables',
the floats are promoted to double and all the math is done in
double-precision. In such cases, the compiler generates floating
point code using emulation library calls, instead of cortex-M4
FPU instructions. To ensure that the FPU instructions are generated,
all the math needs to be done in single-precision. To ensure this,
implicit promotions of floats to doubles must be avoided. Thus,
floating-point literals need to have the 'f' suffix.

Signed-off-by: J. German Rivera <German.Rivera@freescale.com>