Lower the priority of the root system thread after initialization
The root system thread is created with the highest priority
so that it is not preempted by any other thread while doing
system initialization. However, once system initialization is complete,
its priority should be lowered to be just above the idle thread,
so that it does not compete for CPU with time-critical threads.

Signed-off-by: J. German Rivera <German.Rivera@freescale.com>